requirements.txt
alembic==1.0.7
Click==7.0

Flask-Migrate==2.3.1
Flask-Script==2.0.6

Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0

python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
WTForms==2.2.1
flask
Flask-SQLAlchemy
psycopg2

Flask-JWT==0.3.2
PyJWT==1.4.2

